4) LANGCHAIN/PROJECTS/ATS RESUME ANALYZER/requirements.txt
fastapi
langchain
langchain_community
langchain_core
langchain_huggingface
langserve
python-dotenv
uvicorn
pydantic
sse_starlette